fix the failing linter and jest tests that's related to the reconfigured proteinpaint bundling
